% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboRx-package.R
\docType{package}
\name{MetaboRx-package}
\alias{MetaboRx}
\alias{MetaboRx-package}
\title{MetaboRx: QC-Based Signal Drift Correction and Visualization for LC–MS Metabolomics}
\description{
Shiny app to clean, correct, visualize, and export LC-MS metabolomics data using QC-based signal drift correction with batch-wise options.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/breguppy/MetaboRx}
  \item Report bugs at \url{https://github.com/breguppy/MetaboRx/issues}
}

}
\author{
\strong{Maintainer}: Breanna Guppy \email{breanna-guppy@uiowa.edu}

Authors:
\itemize{
  \item Breanna Guppy \email{breanna-guppy@uiowa.edu}
}

}
\keyword{internal}
