% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_required_dependencies.R
\name{check_required_dependencies}
\alias{check_required_dependencies}
\title{Check required QCcorrection dependencies}
\usage{
check_required_dependencies()
}
\value{
Invisibly returns a list of missing require packages
}
\description{
Checks that packages required by QCcorrection are installed
}
\details{
This function checks that core package dependencies listed in \code{Imports}
are installed. Core dependencies are installed automatically when
QCcorrection itself is installed.
}
\examples{
\dontrun{
QCcorrection::check_required_dependencies()
}
}
